refactor(wix-vibe-headless): move the /register repair stub into a real file
The previous commit embedded the Register.jsx page source as a template string
inside deploy.cjs. Page code doesn't belong in the installer — it can't be
reviewed, linted, or edited like the rest of the vertical's pages.

Move it to references/members/install/Register.jsx and let deploy.cjs just copy
it, so both repairs are now plain file copies driven by one path table.

It lives under members/install/ rather than members/app/ on purpose: everything
in app/ is copied into src/ for EVERY members install, and this vertical ships
no /register route by design (sign-up is the "Sign up" tab inside LoginForm;
the routes are /login, /callback, /account). The stub exists solely to
neutralize Base44's seeded Register.jsx, so a fresh app with no such leftover
must not receive it.

Re-verified the same four scenarios: leftovers replaced; re-run idempotent;
an agent's customization survives; a fresh app gets no invented Register.jsx.
Also confirmed no false positive from the Wix Login.jsx header, which mentions
"base44 design tokens" — detection matches only @/api/base44Client and
base44.auth., so the re-run correctly reports left_as_is.
